requirements.txt.txt
Flask==2.3.3
face_recognition==1.3.0
numpy==2.2.6
requests==2.31.0
Pillow==11.3.0
